move setup code
